[move] YourProfile 폴더 이동 #180
