+++ src/pages/Profile/YourProfile/YourProfile.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
-import Layout from '../../components/layout/Layout';
-import OtherProfile from '../../components/common/Profile/OtherProfile';
-import ProfileProduct from '../../components/common/Profile/ProfileProduct';
-import ProfilePost from '../../components/common/Profile/ProfilePost';
-import Modal from '../../components/common/Modal/Modal';
+import Layout from '../../../components/layout/Layout';
+import OtherProfile from '../../../components/common/Profile/OtherProfile';
+import ProfileProduct from '../../../components/common/Profile/ProfileProduct';
+import ProfilePost from '../../../components/common/Profile/ProfilePost';
+import Modal from '../../../components/common/Modal/Modal';
 
 export default function YourProfile() {
   const [isModalShow, setIsModalShow] = useState(false);
